fix: gate the authAxios request interceptor on session teardown (#1392)
authAxios registers its own request interceptor and reads the token
straight from secure storage, with no Redux check at all. It shares the
same 401 handler, so it was exposed to the same re-attachment race: a
request issued mid-teardown read the token being deleted and re-attached
it, earning another 401. The wellness hooks and BreathingTool use this
instance, so the path is live.

Move the teardown state into lib/api/sessionTeardown.ts so both axios
instances gate on one shared promise. The shared module also avoids an
import cycle, since authAxios cannot import from setupAxiosInterceptor,
which already imports authAxios.

No behaviour change for the global instance — teardownExpiredSession()
and the request gate are the same logic, now imported rather than
defined locally.

Adds a test asserting authAxios omits the Authorization header for a
request issued mid-teardown; it re-attaches "Bearer stale-token" without
the gate.
